requirements.txt
pytorch-lightning==2.0.0
tensorboard>=2.14.1
matplotlib>=3.4.3
seaborn>=0.11.2
tqdm>=4.62.3
numpy>=1.16.4
scipy>=1.1.0
SoundFile>=0.10.2
scikit-learn>=0.21.3
pesq
pystoi
librosa
omegaconf
torchmetrics
betacal
h5py
sortedcontainers
onnxruntime
fvcore